advisor/backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
psycopg2-binary==2.9.10
python-json-logger==3.2.1
pydantic-settings==2.7.1
python-nmap==0.7.1
mac-vendor-lookup==0.1.12
alembic==1.14.1
pytest-asyncio==0.25.3
aiosqlite==0.20.0
httpx==0.28.1
